DTOs e Relacionamentos
Finalização da implementação dos DTOs e Inclusão dos relacionamentos ManyToOne

+++ src/main/java/br/unitins/topicos1/dto/AcessoriosDTO.java
@@ -9,8 +9,8 @@ public record AcessoriosDTO(
     Double preco,
     TipoAcessorio acessorio,
     Compatibilidade compatibilidade,
-    Double altura,
-    Double largura,
+    Integer altura,
+    Integer largura,
     String material,
     Double peso,
     Cor cor
